# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NyumbaLink API
  version: "1.0.0"
  description: |
    ## Kenya's Property Marketplace REST API

    **Base URL:** `https://nyumbalink.co.ke/api`

    ### Authentication
    All protected endpoints require a JWT Bearer token obtained from `/auth/login`.
    Include it as: `Authorization: Bearer <access_token>`

    ### Rate Limiting
    | Scope           | Limit             |
    |-----------------|-------------------|
    | Global          | 200 req/day, 50/hr |
    | Auth endpoints  | 10 req/min        |
    | File uploads    | 20 req/hr         |

    ### Pagination
    List endpoints accept:
    | Param     | Type    | Default | Description              |
    |-----------|---------|---------|--------------------------|
    | page      | integer | 1       | Page number (1-indexed)  |
    | per_page  | integer | 12      | Items per page (max 50)  |
    | sort_by   | string  | created_at | Column to sort by     |
    | sort_dir  | string  | desc    | `asc` or `desc`          |

  contact:
    name: NyumbaLink Support
    email: support@nyumbalink.co.ke
    url: https://nyumbalink.co.ke

  license:
    name: Proprietary
    url: https://nyumbalink.co.ke/terms
servers:
  - url: https://nyumbalink.co.ke/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# ──────────────────────────────────────────────────────────────────────
#  Reusable schemas
# ──────────────────────────────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:     { type: integer, example: 248 }
        page:      { type: integer, example: 1 }
        per_page:  { type: integer, example: 12 }
        pages:     { type: integer, example: 21 }
        has_prev:  { type: boolean, example: false }
        has_next:  { type: boolean, example: true }
        sort_by:   { type: string,  example: created_at }
        sort_dir:  { type: string,  enum: [asc, desc], example: desc }
        prev_url:  { type: string,  nullable: true }
        next_url:  { type: string,  nullable: true }
    Error:
      type: object
      required: [error]
      properties:
        error:   { type: string, example: "Resource not found" }
        details: { type: object, description: "Field-level validation errors" }

    # ── User ──────────────────────────────────────────────────────────
    UserPublic:
      type: object
      properties:
        id:         { type: string, format: uuid }
        name:       { type: string, example: "Jane Wanjiru" }
        role:       { type: string, enum: [admin, owner, seeker, agent] }
        avatar_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    UserPrivate:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          properties:
            email:          { type: string, format: email }
            phone:          { type: string, example: "+254712345678", nullable: true }
            email_verified: { type: boolean }
            is_active:      { type: boolean }

    # ── Property ──────────────────────────────────────────────────────
    PropertyImage:
      type: object
      properties:
        id:         { type: string, format: uuid }
        image_url:  { type: string, format: uri }
        is_primary: { type: boolean }
        sort_order: { type: integer }
    Property:
      type: object
      properties:
        id:                  { type: string, format: uuid }
        title:               { type: string, example: "3 Bed Apartment – Kilimani" }
        description:         { type: string, nullable: true }
        category:            { type: string, enum: [rent, short_stay, plot_sale] }
        price:               { type: number, format: float, example: 45000 }
        price_period:        { type: string, enum: [monthly, per_night, yearly, total], nullable: true }
        location:            { type: string, example: "Kilimani, Nairobi" }
        county:              { type: string, example: "Nairobi" }
        town:                { type: string, example: "Kilimani" }
        latitude:            { type: number, nullable: true }
        longitude:           { type: number, nullable: true }
        bedrooms:            { type: integer, nullable: true }
        bathrooms:           { type: integer, nullable: true }
        plot_size:           { type: number, nullable: true }
        plot_size_unit:      { type: string, nullable: true }
        floor_area:          { type: number, nullable: true }
        amenities:           { type: array, items: { type: string } }
        owner_id:            { type: string, format: uuid }
        verification_status: { type: string, enum: [pending, verified, rejected] }
        status:              { type: string, enum: [active, inactive, sold, rented] }
        featured:            { type: boolean }
        views_count:         { type: integer }
        images:              { type: array, items: { $ref: '#/components/schemas/PropertyImage' } }
        primary_image:       { $ref: '#/components/schemas/PropertyImage', nullable: true }
        avg_rating:          { type: number, nullable: true }
        created_at:          { type: string, format: date-time }
        updated_at:          { type: string, format: date-time }
    PropertyCreate:
      type: object
      required: [title, category, price, location]
      properties:
        title:        { type: string, minLength: 5, maxLength: 500 }
        description:  { type: string, maxLength: 5000 }
        category:     { type: string, enum: [rent, short_stay, plot_sale] }
        price:        { type: number, minimum: 0 }
        price_period: { type: string, enum: [monthly, per_night, yearly, total] }
        location:     { type: string }
        county:       { type: string }
        town:         { type: string }
        latitude:     { type: number }
        longitude:    { type: number }
        bedrooms:     { type: integer, minimum: 0 }
        bathrooms:    { type: integer, minimum: 0 }
        plot_size:    { type: number, minimum: 0 }
        floor_area:   { type: number, minimum: 0 }
        amenities:    { type: array, items: { type: string } }

    # ── Inquiry ───────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /auth/register:
    post:
      summary: Create a new user account
      operationId: register
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:     { type: string, example: "Jane Wanjiru" }
                email:    { type: string, format: email }
                password: { type: string, minLength: 8 }
                phone:    { type: string, example: "+254712345678" }
                role:     { type: string, enum: [seeker, owner], default: seeker }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:          { $ref: '#/components/schemas/UserPrivate' }
                  access_token:  { type: string }
                  refresh_token: { type: string }
        "409": { description: Email already registered }
        "400": { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      summary: Log in and get JWT tokens
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:          { $ref: '#/components/schemas/UserPrivate' }
                  access_token:  { type: string }
                  refresh_token: { type: string }
        "401": { description: Invalid credentials }
  /auth/refresh:
    post:
      summary: Refresh access token using refresh token
      operationId: refreshToken
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
  /auth/me:
    get:
      summary: Get current user profile
      operationId: getMe
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/UserPrivate' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    put:
      summary: Update profile (name, phone)
      operationId: updateMe
      tags: [Auth]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:  { type: string }
                phone: { type: string }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/UserPrivate' }
  /auth/logout:
    post:
      summary: Revoke tokens and log out
      operationId: logout
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out

  # ── Properties ────────────────────────────────────────────────────
  /properties:
    get:
      summary: List published properties (with filters + pagination)
      operationId: listProperties
      tags: [Properties]
      parameters:
        - { name: page,        in: query, schema: { type: integer, default: 1 } }
        - { name: per_page,    in: query, schema: { type: integer, default: 12, maximum: 50 } }
        - { name: sort_by,     in: query, schema: { type: string, enum: [price, created_at, views_count, bedrooms] } }
        - { name: sort_dir,    in: query, schema: { type: string, enum: [asc, desc], default: desc } }
        - { name: category,    in: query, schema: { type: string, enum: [rent, short_stay, plot_sale] } }
        - { name: county,      in: query, schema: { type: string }, example: Nairobi }
        - { name: town,        in: query, schema: { type: string }, example: Kilimani }
        - { name: min_price,   in: query, schema: { type: number } }
        - { name: max_price,   in: query, schema: { type: number } }
        - { name: bedrooms,    in: query, schema: { type: integer } }
        - { name: min_plot,    in: query, schema: { type: number } }
        - { name: q,           in: query, schema: { type: string }, description: Keyword search }
        - { name: featured,    in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Paginated property list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: '#/components/schemas/Property' } }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }

    post:
      summary: Create a new property listing (owners only)
      operationId: createProperty
      tags: [Properties]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyCreate' }
      responses:
        "201":
          description: Property created (status=pending verification)
          content:
            application/json:
              schema:
                type: object
                properties:
                  property: { $ref: '#/components/schemas/Property' }
                  message:  { type: string }
        "400": { $ref: '#/components/responses/ValidationError' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /properties/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      summary: Get property details (public)
      operationId: getProperty
      tags: [Properties]
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        "404": { $ref: '#/components/responses/NotFound' }

    put:
      summary: Update property (owner or admin)
      operationId: updateProperty
      tags: [Properties]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyCreate' }
      responses:
        "200": { description: Updated property }
        "403": { $ref: '#/components/responses/Forbidden' }

    delete:
      summary: Delete property (owner or admin)
      operationId: deleteProperty
      tags: [Properties]
      security:
        - BearerAuth: []
      responses:
        "200": { description: Deleted }
        "403": { $ref: '#/components/responses/Forbidden' }
...
  /admin/dashboard:
    get:
      summary: Admin dashboard statistics
      operationId: adminDashboard
      tags: [Admin]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Platform-wide stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:      { type: object }
                  properties: { type: object }
                  payments:   { type: object }
        "403": { $ref: '#/components/responses/Forbidden' }
  /admin/properties/{id}/verify:
    put:
      summary: Approve or reject a property listing
      operationId: verifyProperty
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verification_status]
              properties:
                verification_status: { type: string, enum: [verified, rejected, pending] }
                reason:              { type: string, description: Required when rejecting }
      responses:
        "200": { description: Verification status updated }
  /admin/users:
    get:
      summary: List all users
      operationId: adminListUsers
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - { name: role,     in: query, schema: { type: string } }
        - { name: q,        in: query, schema: { type: string } }
        - { name: page,     in: query, schema: { type: integer } }
        - { name: per_page, in: query, schema: { type: integer } }
      responses:
        "200": { description: Paginated users }

  # ── SEO ───────────────────────────────────────────────────────────
...